Combine four Python tests into one
To reduce the test time
