+++ src/workerd/server/tests/python/BUILD.bazel
@@ -3,14 +3,13 @@ load("//src/workerd/server/tests/python:py_wd_test.bzl", "py_wd_test", "python_t
 
 python_test_setup()
 
+# Each test takes a while to start up, so if possible add additional tests to many-tests or sdk or
+# some other existing test.
 py_wd_test("hello")
 
-py_wd_test("env-param")
+py_wd_test("many-tests")
 
-py_wd_test(
-    "top-level-env",
-    skip_python_flags = ["0.26.0a2"],
-)
+py_wd_test("env-param")
 
 py_wd_test("asgi")
 
@@ -21,8 +20,6 @@ py_wd_test("random")
 py_wd_test("subdirectory")
 
 py_wd_test("sdk")
-
-py_wd_test("seek-metadatafs")
 
 gen_rust_import_tests()
 
@@ -50,17 +47,11 @@ py_wd_test(
     skip_python_flags = ["0.26.0a2"],
 )
 
-py_wd_test("js-import")
-
-py_wd_test("importable-env")
-
 py_wd_test("python-rpc")
 
 py_wd_test("workflow-entrypoint")
 
 py_wd_test("vendor_dir_compat_flag")
-
-py_wd_test("multiprocessing")
 
 py_wd_test("default-class-with-legacy-global-handlers")
