--- src/workerd/server/tests/python/js-import/js-import.wd-test
@@ -1,18 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "js-import-test",
-      worker = (
-        modules = [
-          (name = "worker.py", pythonModule = embed "worker.py"),
-          (name = "child", esModule = "export const value = 'Child module loaded successfully';"),
-        ],
-        compatibilityFlags = [%PYTHON_FEATURE_FLAGS, "disable_python_no_global_handlers"],
-        bindings = [
-          (name = "TEST_VALUE", text = "TEST_STRING"),
-        ],
-      )
-    ),
-  ],
-);
