+++ src/workerd/server/tests/python/many-tests/env.wd-test
@@ -2,18 +2,23 @@ using Workerd = import "/workerd/workerd.capnp";
 
 const unitTests :Workerd.Config = (
   services = [
-    ( name = "importable-env-test",
+    ( name = "python-hello",
       worker = (
         modules = [
-          (name = "worker.py", pythonModule = embed "worker.py"),
+          (name = "worker.py", pythonModule = embed "worker.py")
         ],
         compatibilityFlags = [
           %PYTHON_FEATURE_FLAGS,
-          "disable_python_no_global_handlers",
+          "python_no_global_handlers",
           "unwrap_custom_thenables"
         ],
         bindings = [
+          (
+            name = "secret",
+            text = "thisisasecret"
+          ),
           (name = "FOO", text = "BAR"),
+          (name = "TEST_VALUE", text = "TEST_STRING"),
           (name = "CACHE", memoryCache = (
             id = "abc123",
             limits = (
