--- src/workerd/server/tests/python/multiprocessing/multiprocessing.wd-test
@@ -1,14 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "python-hello",
-      worker = (
-        modules = [
-          (name = "worker.py", pythonModule = embed "worker.py")
-        ],
-        compatibilityFlags = [%PYTHON_FEATURE_FLAGS, "disable_python_no_global_handlers", "disable_python_check_rng_state"],
-      )
-    ),
-  ],
-);
