--- src/workerd/server/tests/python/multiprocessing/worker.py
@@ -1,9 +1,0 @@
-import multiprocessing
-
-
-def use(x):
-    pass
-
-
-def test():
-    use(multiprocessing)
