--- src/workerd/server/tests/python/seek-metadatafs/seek-metadatafs.wd-test
@@ -1,15 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "main",
-      worker = (
-        modules = [
-          (name = "a.py", pythonModule = embed "./a.py"),
-        ],
-        compatibilityFlags = [%PYTHON_FEATURE_FLAGS, "disable_python_no_global_handlers"],
-      )
-    ),
-  ],
-);
-
