--- src/workerd/server/tests/python/top-level-env/env.wd-test
@@ -1,20 +1,0 @@
-using Workerd = import "/workerd/workerd.capnp";
-
-const unitTests :Workerd.Config = (
-  services = [
-    ( name = "python-hello",
-      worker = (
-        modules = [
-          (name = "worker.py", pythonModule = embed "worker.py")
-        ],
-        bindings = [
-          (
-            name = "secret",
-            text = "thisisasecret"
-          ),
-        ],
-        compatibilityFlags = [%PYTHON_FEATURE_FLAGS],
-      )
-    ),
-  ],
-);
